requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
langchain==0.0.332
langchain-google-genai==0.0.5
google-generativeai==0.3.1
supabase==2.0.3
mercadopago==2.2.0
pytest==7.4.3
httpx==0.24.1
jinja2==3.1.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4 